Gerenciamento de Clientes

+++ src/turismo/models/Pessoas/Funcionario.java
@@ -3,7 +3,7 @@ package turismo.models.Pessoas;
 import java.time.LocalDate;
 
 public class Funcionario extends Usuario {
-    public Funcionario(String nome, int documento, int id, int telefone, String email, LocalDate aniversario) {
+    public Funcionario(String nome, long documento, long id, long telefone, String email, LocalDate aniversario) {
         super(nome, documento, id, telefone, email, aniversario);
     }
 }
